migrate helper objects to kotlin extensions
